Remove the dependencies to the generic problem field.
The problem field is deprecated to remain retrocompatible. However, it is
aimed to be removed with other deprecated stuff when judged ready for it.
The point is to remove fields that pollute solutions not needing them.
Problems in particular are mainly used at the instanciation of the solution
while other data, like number of solutions and such, are only required by
some solutions during their lifetime. These solution can store the problem
by themselves or store the required data only. This way, solutions not
needing the problem save some space.